Prod Release

+++ Listrr/Jobs/BackgroundJobs/ProcessMovieListBackgroundJob.cs
@@ -1,4 +1,10 @@
-﻿using Hangfire;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+
+using Hangfire;
 
 using Listrr.Comparer;
 using Listrr.Configuration;
@@ -8,11 +14,8 @@ using Listrr.Jobs.RecurringJobs;
 using Listrr.Repositories;
 using Listrr.Services;
 
-using System;
-using System.Linq;
-using System.Threading.Tasks;
-
 using TraktNet.Exceptions;
+using TraktNet.Objects.Get.Movies;
 
 namespace Listrr.Jobs.BackgroundJobs
 {
@@ -23,7 +26,7 @@ namespace Listrr.Jobs.BackgroundJobs
         private readonly TraktAPIConfiguration _traktApiConfiguration;
 
         private TraktList traktList;
-        
+
         public ProcessMovieListBackgroundJob(ITraktService traktService, TraktAPIConfiguration traktApiConfiguration, ITraktListRepository traktRepository)
         {
             _traktService = traktService;
@@ -39,28 +42,64 @@ namespace Listrr.Jobs.BackgroundJobs
                 traktList = await _traktService.Get(traktList);
 
                 traktList.ScanState = ScanState.Updating;
-                
+
                 await _traktRepository.Update(traktList);
 
-                var found = await _traktService.MovieSearch(traktList);
-                var existing = await _traktService.GetMovies(traktList);
-                
-                var remove = existing.Except(found, new TraktMovieComparer()).ToList();
-                var add = found.Except(existing, new TraktMovieComparer()).ToList();
-
-                if (add.Any())
+                if (string.IsNullOrWhiteSpace(traktList.ItemList))
                 {
-                    foreach (var toAddChunk in add.ChunkBy(_traktApiConfiguration.ChunkBy))
+                    var found = await _traktService.MovieSearch(traktList);
+                    var existing = await _traktService.GetMovies(traktList);
+
+                    var remove = existing.Except(found, new TraktMovieComparer()).ToList();
+                    var add = found.Except(existing, new TraktMovieComparer()).ToList();
+
+                    if (add.Any())
                     {
-                        await _traktService.AddMovies(toAddChunk, traktList);
+                        foreach (var toAddChunk in add.ChunkBy(_traktApiConfiguration.ChunkBy))
+                        {
+                            await _traktService.AddMovies(toAddChunk, traktList);
+                        }
+                    }
+
+                    if (remove.Any())
+                    {
+                        foreach (var toRemoveChunk in remove.ChunkBy(_traktApiConfiguration.ChunkBy))
+                        {
+                            await _traktService.RemoveMovies(toRemoveChunk, traktList);
+                        }
                     }
                 }
-
-                if (remove.Any())
+                else
                 {
-                    foreach (var toRemoveChunk in remove.ChunkBy(_traktApiConfiguration.ChunkBy))
+                    var add = new List<ITraktMovie>();
+                    var regex = new Regex(@"(.*)\(([0-9]{4})\)");
+
+                    foreach (var line in traktList.ItemList.Split("\r\n"))
                     {
-                        await _traktService.RemoveMovies(toRemoveChunk, traktList);
+                        var processedLine = regex.Match(line);
+                        if (processedLine.Success && processedLine.Groups.Count == 3)
+                        {
+                            var cleanMovieName = processedLine.Groups[1].Value.Trim();
+                            var itemYearParseResult = int.TryParse(processedLine.Groups[2].Value, out var movieYear);
+                            if (itemYearParseResult)
+                            {
+                                var itemResult = await _traktService.MovieSearch(traktList, cleanMovieName, movieYear);
+                                if (itemResult != null)
+                                {
+                                    add.Add(itemResult);
+                                }
+
+                                await Task.Delay(_traktApiConfiguration.DelayIdSearch);
+                            }
+                        }
+                    }
+
+                    if (add.Any())
+                    {
+                        foreach (var toAddChunk in add.ChunkBy(_traktApiConfiguration.ChunkBy))
+                        {
+                            await _traktService.AddMovies(toAddChunk, traktList);
+                        }
                     }
                 }
 
@@ -97,15 +136,15 @@ namespace Listrr.Jobs.BackgroundJobs
                 {
                     traktList.ScanState = ScanState.None;
 
-                    if(forceRefresh)
+                    if (forceRefresh)
                         await _traktService.Update(traktList);
 
                     await _traktRepository.Update(traktList);
                 }
             }
 
-            if(queueNext)
-                BackgroundJob.Enqueue<ProcessUserListsRecurringJob>(x => x.Execute());
+            if (queueNext)
+                BackgroundJob.Schedule<ProcessUserListsRecurringJob>(x => x.Execute(), TimeSpan.FromSeconds(_traktApiConfiguration.DelayRequeue));
         }
 
         [Queue("donor")]
